Revert "NN-3037 return cell move reasons (#215)"
This reverts commit 8aa951770f9860d47edaf25cbfc077cdff7cec2e.
