+++ src/test/kotlin/uk/gov/justice/digital/hmpps/whereabouts/integration/CellMoveIntegrationTest.kt
@@ -41,15 +41,4 @@ class CellMoveIntegrationTest : IntegrationTest() {
       )
     )
   }
-
-  @Test
-  fun `return all cell move reasons`() {
-    webTestClient.get()
-      .uri("/cell/cell-move-reasons")
-      .headers(setHeaders())
-      .exchange()
-      .expectStatus().isOk
-      .expectBody()
-      .json(loadJsonFile("cell-move-reasons.json"))
-  }
 }
