--- src/test/resources/uk/gov/justice/digital/hmpps/whereabouts/integration/cell-move-reasons.json
@@ -1,28 +1,0 @@
-{
-  "reasons": [
-    {
-      "code": "ADM",
-      "description": "Administrative"
-    },
-    {
-      "code": "BEH",
-      "description": "Behaviour"
-    },
-    {
-      "code": "CLA",
-      "description": "Classification or re-classification"
-    },
-    {
-      "code": "CON",
-      "description": "Conflict with other prisoners"
-    },
-    {
-      "code": "LN",
-      "description": "Local needs"
-    },
-    {
-      "code": "VP",
-      "description": "Vulnerable prisoner"
-    }
-  ]
-}
